+++ src/main/java/de/samply/token/DataShieldTokenManagerJob.java
@@ -6,7 +6,8 @@ import de.samply.db.repository.ProjectBridgeheadUserRepository;
 import de.samply.project.ProjectType;
 import de.samply.project.state.ProjectState;
 import de.samply.token.dto.DataShieldProjectStatus;
-import de.samply.token.dto.DataShieldStatus;
+import de.samply.token.dto.DataShieldTokenManagerProjectStatus;
+import de.samply.token.dto.DataShieldTokenManagerTokenStatus;
 import de.samply.token.dto.DataShieldTokenStatus;
 import de.samply.user.roles.ProjectRole;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -42,11 +43,11 @@ public class DataShieldTokenManagerJob {
         activeUsers.addAll(this.projectBridgeheadUserRepository.getByProjectTypeAndProjectStateAndProjectRole(ProjectType.DATASHIELD, ProjectState.FINAL, ProjectRole.FINAL));
         activeUsers.forEach(user -> {
             // Check user status
-            DataShieldStatus dataShieldStatus = tokenManagerService.fetchProjectStatus(user.getProjectBridgehead().getProject().getCode(), user.getProjectBridgehead().getBridgehead());
-            if (dataShieldStatus.projectStatus() == DataShieldProjectStatus.WITH_DATA) {
-                if (dataShieldStatus.tokenStatus() == DataShieldTokenStatus.NOT_FOUND) { // If user token not found: Create token
+            DataShieldTokenManagerTokenStatus dataShieldTokenManagerTokenStatus = tokenManagerService.fetchTokenStatus(user.getProjectBridgehead().getProject().getCode(), user.getProjectBridgehead().getBridgehead(), user.getEmail());
+            if (dataShieldTokenManagerTokenStatus.projectStatus() == DataShieldProjectStatus.WITH_DATA) {
+                if (dataShieldTokenManagerTokenStatus.tokenStatus() == DataShieldTokenStatus.NOT_FOUND) { // If user token not found: Create token
                     tokenManagerService.generateTokensInOpal(user.getProjectBridgehead().getProject().getCode(), user.getProjectBridgehead().getBridgehead(), user.getEmail());
-                } else if (dataShieldStatus.tokenStatus() == DataShieldTokenStatus.EXPIRED) { // If user token expired: Refresh Token
+                } else if (dataShieldTokenManagerTokenStatus.tokenStatus() == DataShieldTokenStatus.EXPIRED) { // If user token expired: Refresh Token
                     tokenManagerService.refreshToken(user.getProjectBridgehead().getProject().getCode(), user.getProjectBridgehead().getBridgehead(), user.getEmail());
                 }
             }
@@ -60,9 +61,9 @@ public class DataShieldTokenManagerJob {
         inactiveUsers.addAll(this.projectBridgeheadUserRepository.getByProjectTypeAndProjectStateAndNotProjectRole(ProjectType.DATASHIELD, ProjectState.FINAL, ProjectRole.FINAL));
         inactiveUsers.forEach(user -> {
             // Check user status
-            DataShieldStatus dataShieldStatus = tokenManagerService.fetchProjectStatus(user.getProjectBridgehead().getProject().getCode(), user.getProjectBridgehead().getBridgehead());
+            DataShieldTokenManagerTokenStatus dataShieldTokenManagerTokenStatus = tokenManagerService.fetchTokenStatus(user.getProjectBridgehead().getProject().getCode(), user.getProjectBridgehead().getBridgehead(), user.getEmail());
             // If user token created or expired: Remove token
-            if (dataShieldStatus.tokenStatus() != DataShieldTokenStatus.NOT_FOUND) {
+            if (dataShieldTokenManagerTokenStatus.tokenStatus() != DataShieldTokenStatus.NOT_FOUND) {
                 tokenManagerService.removeTokens(user.getProjectBridgehead().getProject().getCode(), user.getProjectBridgehead().getBridgehead(), user.getEmail());
             }
         });
@@ -76,9 +77,9 @@ public class DataShieldTokenManagerJob {
             if (!removedProjects.contains(user.getProjectBridgehead().getProject().getCode())) {
                 removedProjects.add(user.getProjectBridgehead().getProject().getCode());
                 // Check user status
-                DataShieldStatus dataShieldStatus = tokenManagerService.fetchProjectStatus(user.getProjectBridgehead().getProject().getCode(), user.getProjectBridgehead().getBridgehead());
+                DataShieldTokenManagerProjectStatus dataShieldTokenManagerProjectStatus = tokenManagerService.fetchProjectStatus(user.getProjectBridgehead().getProject().getCode(), user.getProjectBridgehead().getBridgehead());
                 // if project not found: Remove Token and project
-                if (dataShieldStatus.projectStatus() != DataShieldProjectStatus.NOT_FOUND) {
+                if (dataShieldTokenManagerProjectStatus.projectStatus() != DataShieldProjectStatus.NOT_FOUND) {
                     tokenManagerService.removeProjectAndTokens(user.getProjectBridgehead().getProject().getCode(), user.getProjectBridgehead().getBridgehead());
                 }
             }