+++ src/main/java/de/samply/token/dto/DataShieldTokenManagerProjectStatus.java
@@ -1,0 +1,10 @@
+package de.samply.token.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record DataShieldTokenManagerProjectStatus(
+        @JsonProperty("project_id") String projectCode,
+        @JsonProperty("bk") String bridgehead,
+        @JsonProperty("project_status") DataShieldProjectStatus projectStatus
+        ) {
+}